mapstore: add get_resources_using_ows
returns a set of tuples (restype, id) using a given ows resource.
the ows resource can be a wxs endpoint, or a layer in an endpoint.

currently super expensive as it builds reverse maps at each calls because:
- i haven't looked yet how to properly cache/update that
- nor how to filter sql on json fields, since gs_stored_data is a
varchar (instead of a proper json-typed field) and mapstore stores seemingly
invalid json in it...