ows: add an optional single arg to owslayer() check
and pass it to owscache.get() to eventually force-fetch the capabilities cache.
the single arg defaults to false.

checking a single owslayer should force-fetch the capabilities document to make
sure we work on the very latest layer description, but not as part of an
owsservice full check, which will refresh the capabilities cache once at the
start.