csw: add an optional single arg to check_record()
like done in commit c011ada for ows layer.

pass it to owscache.get() to eventually force-fetch the capabilities cache.
the single arg defaults to false.

checking a single csw record should force-fetch the capabilities document to
make sure we work on the very latest csw record description, but not as part of
an cswservice full check, which will refresh the capabilities cache once at the
start.

actually, checking a single csw record should only refresh our view of the
given record....