script: split/rewrite fetchForHome
now it mostly calls /tasks/lastresultbytask/ from #4 to show latest
task results, and update the corresponding bootstrap card from 6d8108a43

if the call was successful, display the amount of items, the amount of
errors and the timestamp of the task.
show a warning if no task was run/found for the given taskname/args couple.

the bit calling /api/geonetwork/metadatas.json is splitted in
fetchMyMd method (called from /my-metadata page) and the bits calling
/api/mapstore/{maps,contexts}.json are splitted to fetchMyMaps
(called from /my-maps-and-apps page)