redisbackend: return finished as a posix timestamp in get_taskids_by_taskname_and_args() (#9)
the callers of get_taskids_by_taskname_and_args() all pass the value to the
jinja tojson() filter, which doesnt know how to deal with datetime objects, so
we end up with an ugly 'date-as-string'.

using a timestamp/numeric value allows the js frontend to build a proper Date()
object, thus allowing nicer formatting.